Added code to TreeConvert to read INGRID files.
Added code to the tree convert code to read the INGRID selection files
from Benjamin. It's not complete as Benjamin is going to add a couple
more variables. But it has enough information to test the fit.

Added/edited binning files for INGRID. As the INGRID analysis uses
different reconstructed and true variables, there is a binning file for
each one.