ND280+INGRID template/flux/detector fit.
Slight updates/modifications for fitting ND280+INGRID with template,
flux, and detector parameters.

New script to create a dummy INGRID detector covariance and merge it
with the ND280 detector covariance.

Small addition to OptParser to include an option to apply the POT
normalization to a given sample or not.

Also committing before more hacks are made to add in the cross-section
parameters.