Enabled xsec parameters. Updates to inputs.
Enabled the cross-section parameters for INGRID with an ugly hack. This
hack skips defining Minuit parameters for the INGRID dials so that the
ND280 dials take care of both detectors/splines.

This needs to be improved, or at least made into a config option.

Updates to input files, especially the binning files for INGRID. They
needed the header line to declare how many dimensions to use for
BinManager.