ND280+INGRID error propagation running.
With a few configuration file tweaks the ND280+INGRID error propagation
code works.

When defining signals using sample number, the truth samples need to
also be included.

Slight edit to FitObj to include the POT scaling option. This will
further need to be edited as the normalization is still applied to all
signal histograms when reweighting.