Fixes to errprop flux hist reweighting. Fix to TreeConvert.
Mainly fixes to the flux hist reweighting in the error propagation code.
The correct offset/index was not being calculated correctly for the flux
parameter vector. The detector offset was not being included.

The TreeConverter code was still saving enu for INGRID in GeV, which was
then still being divided by 1000 in FluxParameters. This hardcoding
really needs to be removed.

Updates to configuration files.

Finally changed "iterations" to "func calls" to be more accurate with
respect to Minuit terminology.