Updates to perform fake data tests.
Updated INGRID analysis binning to help with bins having zero second
derivatives in the fit. Mainly the last two bins have been combined into
a single 1-30 GeV bin.

Updated the binning files in the config.

New flag for the tree converter to enable/disable extra event weights.
Also added a function to provide weights based on enu.

Updated how the POT normalization is done in ErrProp to handle fake data
plots. Originally the POT normalization was being applied to the "data",
which is incorrect. There is a better solution, but for now a new
function argument was added to check whether to apply the POT norm.