New code to produce the INGRID detector cov in reco bins.
Wrote new code to make the INGRID detector cov in reco bins from the
covariance in true bins.

It is run with a small JSON config and a couple binning files.

The errors looks somewhat large compared to the values in the technote,
but this could be because of binning effects.