Minor updates to plotting. Changed to use SVD decomposition.
Changed the PCA to use SVD decomposition instead of eigen decomposition.

Minor updates to the plotting script to fix the labelling for fit
parameter plots, and add TLine's to separate ND280 and INGRID bins.

Updated submit script to include a directory for the results to be
placed in. Also commiting the fit job file itself, only meaningful
change is the job time -- 12 hours.