Added function to get effective number of events in a sample.
Added a function to AnaSample to return the effective number of events
in the sample prediction histogram. This simply returns the integral of
the bins (excluding under/over flow bins).

Realized the current direction of the implementation is not sufficient.
Further commits will change the design of how to constrain the extra
parameters.