Implemented constrained parameters using different method.
Implemented the constraint on the extra parameters using a different
method. Now AnaFitParameters has a virtual function to calculate a
constraint given the sample vector and parameters.

If this function is not defined in a derived class it simply returns the
parameter vector unaltered. This is called for each fit parameter class
in XsecFitter during the minimization.

This appears to be working, however the extra parameters are difficult
to minimize without issue. Only template+extra asimov fits work
completely (e.g. no forcing positive definite) at the moment.