Fixed implementation of ExtraParameters in ErrorProp
The addition of the constraint on the extra parameters to preserve the
number of events required a few more changes to the error propagation to
handle it correctly.

Most notably the function to calculate the constraint needed to be
added, and not calculated for the truth tree.

Also the sample histograms now need to be filled to calculate the number
of events per sample. Although this could be changed to simply multiply
all the event weights together rather than using the integral of the
histograms.