Minor edits to code. Updating xsec covariance.
A couple minor edits relating to debugging the extra migration
parameters. Added an abs() function to determine when to output if the
extra parameters are different from one.

Added some code to change the weights of events to see if the extra
parameters could be fit to a known variation.

Updated normalized xsec covariance.