Removed GetSampleBreakdown()
Removed the GetSampleBreakdown function from AnaSample().

This function produced a summary of the topology distribution for each
sample and some histograms.

However it relied on a vector of strings to label each topology (and
inform the function of the total number of options) which was only used
in this one function and had to be specified in the config file.

This config setting was confusing to its purpose and the function itself
only serves to be a minor diagonistic, which is probably mostly ignored
in the log output.

It has been removed along with this config setting for now and this
functionality will eventually be replaced elsewhere.