Removing deprecated code and histograms
Removed the commented/deprecated code in the FillEventHist() loop which
was recalculating the bin index every time the function was called.

Removed all other histograms other than the prediction and data
histograms. These other histograms are never used anywhere in the code
except when writing to output, and even then are not that useful and may
even be completely useless (for instance if the true and reco sample
binning are not the same).

Other minor changes:
- Added some const correctness to a couple variables/functions
- Changd some stdout statements to use the TAGS
- Removed Get/Set functions for the above histograms