AnaSample now uses the vector of fit variables
AnaSample now uses the vector of the kinematic fit variables contained
in the AnaEvent. This vector is directly passed to the BinManager to
calculate the bin index.

Needed to change the variable type from float to double in all places as
the BinManager specifies a std::vector<double> in the function
signature. (I should just add a <float> version of this function...)