Modified FitParameters to use the BinManager
Changed the FitParameter class to use the BinManager for handling the
binning instead of the original GetBinIndex function.

This also now allows the FitParameters to use an N-dimensional binning,
and InitEventMap() now uses the true variable vector from the event
class.

Deleted now obsolete functions and variables caused by the change.