Split FillEventHist function into two functions
Split the AnaSample::FillEventHist function in to two functions
to handle filling the data and MC histograms separately.

FillEventHist now only fills the MC prediction histogram and takes one
parameter, a boolean on whether to use the current event weight or the
nominal MC weight.

The new function, FillDataHist, now only fills the data histograma and
still takes the same parameters, i.e. fill data with the MC or from a
file and whether or not to apply statistical fluctuations.

XsecFitter has been edited accordingly to use the appropriate functions
and arguments.