Update DetParameters to use the BinManager
Updated the detector parameters class to use a BinManager for the
binning for each sample.

Modified the InitEventMap to use the reco variable vector from the
AnaEvent allowing for N dimensions to be used for detector binning.

Removed obsolete functions. Changed AddDetector function signature to
remove the bool flag as it was always hardcoded to be true.

Modified function name to get the binning file from AnaSample.

Corresponding edits in FitObj and xsllhFit to update function calls.