Removed options to set a global POT ratio.
Removed the options and functions to set a global POT ratio for all
events. This is not the correct method to set the Data/MC POT ratio, and
does not work for analyses working with datasets needing different POT
ratios.

For now the AnaSample still has the function to set an overall sample
normalization which could be used for a per-sample basis POT ratio.
However there are no configuration options to use this.

Ideally the POT ratio should be applied to the events during the tree
convert stage.