Additional AnaEvent variable
Added the beam mode (horn current) as a variable to the AnaEvent and
associated Set/Get functions. Events now carry information if the event
occurred in nu-mode running or antinu-mode running (or FHC/RHC).

Adjusted the default values for several AnaEvent variables to -99 rather
than -1. This is because some analyzers might want to use -1 as a valid
value (particularly for the beam mode). A value of -99 should be a more
safe default that is unlikely to collide with a valid value.

Minor edits to FitParameters and XsecParameters to remove now
unnecessary header files.