Remove obsolete config option and rename 'cut_branch'
Removed the 'binning' config option from the 'detectors' array as it is
never used in the current version of the fit code. This only affected
the OptParser code (and a commented out line in xsllhFit).

Renamed 'cut_branch' to 'sample_id' for the 'sample' definitions in the
config file and in OptParser for better clairty. 'cut_branch' is a bit
of T2K jargon from HL and 'sample_id' is what the value is being used
for.