Rewrite README and minor deletions of unused code
Rewrote the README to remove the even more outdated instructions on how
to install and use the fit code. README now refers to the online and in
repository documenation which is slightly more up-to-date.

Removed last remaining references to the FitStructs header, and deleted
the header file itself.

Removed xsllhUnitTests module as it was never used. Some sort of unit
test suite needs to be developed... another one for "If I can find the
time..."

Modified the src/CMakeLists.txt to remove the xsllhUnitTests executable.

Whitespace/formatting changes to include/ProgressBar.hh